correctly read semi-transparent polys from model
the polygon list for an object consists of four arrays:
- semi-transparent quads
- opaque quads
- semi-transparent tris
- opaque tris

most models do not use any semi-transparent polys, so they have zero length
arrays before the opaque quad and opaque tri lists. these empty arrays are
represented by the byte string `00000000`. i didn't understand what this was
there for, and had been skipping it automatically, which would crash the ripper
if it tried to load a model that had a non empty array there