export scale keyframes
scale seems to be rarely used, but a few digimon have geometry that is not meant
to be visible in some animations, e.g. digitamamon's cracked shell. these are
hidden by giving a zero transformation matrix. so we need to decompose the
matrix for each keyframe into a rotation and a scale, instead of assuming the
matrix is just a rotation